requirements.txt
python-dotenv
nltk
tweepy
pandas
gensim

matplotlib
wordcloud